Move bankroll library under its own folder
