+++ bankroll/csvsectionslicer.py
