+++ bankroll/model/instrument.py
