+++ bankroll/model/marketdata.py
