+++ tests/helpers.py
@@ -1,7 +1,7 @@
 from datetime import date, datetime
 from decimal import Decimal
 from hypothesis.strategies import builds, dates, datetimes, decimals, from_regex, from_type, just, lists, integers, none, one_of, register_type_strategy, sampled_from, text, SearchStrategy
-from model import Activity, Cash, Currency, Instrument, Stock, Bond, Option, OptionType, FutureOption, Future, Forex, Position, DividendPayment, Trade, TradeFlags, Quote
+from bankroll import Activity, Cash, Currency, Instrument, Stock, Bond, Option, OptionType, FutureOption, Future, Forex, Position, DividendPayment, Trade, TradeFlags, Quote
 from typing import List, Optional, TypeVar
 
 T = TypeVar('T')
